Speed up the rsync script.
The trouble is that the rsync is run for every host, whether that host
has any tarballs for copying or not. Unfortunately, that takes about 5
seconds, even if the host directory is empty. For a couple of hundred
hosts, we end up spending 20 minutes in this loop, even if there is
nothing to be done.

Now the script syncs only those hosts which contain new tarballs: if a
host does not contain any new tarballs, it is skipped. Two new state
directories, TO-SYNC and SYNCED, track which hosts contain new
tarballs and therefore need syncing. Prefix files are handled properly
as well.

pbench-base.sh creates TO-SYNC and SYNCED state subdirs. Note that
only tarballs in the main host directory are synced: state subdirs are
excluded.

To avoid multiple ssh invocations for the cleanup, we assume that a
script in the remote can be invoked to DTRT: a single ssh invocation
of that script is all that is necessary. The script itself is part
of a different commit under this PR.

Fixes issues 537, 555(partially - there is still an occasional failure
in unpacking that has not been explained), 557.