Server installation ansible playbooks and roles
[This PR depends on PR#1196.]

These playbooks implement a complete install and setup
of one or more pbench servers.

In addition to installing the pbench-server RPM and the config file
for each server, they do all the work that the pbench-server-activate-*
scripts do.

Most of the scripts haven been "ansiblized": they use appropriate
ansible modules in order to carry out their operations.

The one exception is the pbench-server-activate-create-crontab role:
the ansible script just executes the command remotely.  The reason for
that is that reimplementing requires fetching too many variables from
the config file in the server.

The other scripts already do a fair amount of that (see the
pbench-server-vars role) but the crontab role would double the number
of variables.

This commit makes all the server/bin/pbench-server-activate-* scripts
(*except* for pbench-server-activate-create-crontab) obsolete. They
will be removed in a subsequent PR, once we are sure that the ansible
playbooks are reliable.

The pbench-server-activate-setup-web-server role uses the synchronize
module, a wrapper around rsync, so the role ensures that rsync is
installed.

Since we are synchronizing a directory on each remote host, each
remote host has to be able to login passwordlessly to itself (*not* to
localhost: on each host as root, do "ssh-copy-id root@<host>".)

Change the install state of the RPM from "present" to "latest"