Push the connector object down into the module.
The S3 connector is *very simple*: it implements the methods
that the underlying S3 client provides as pass-through methods.

The mock S3 connector should provide the same methods with the
same interface, but using the filesystem for state.

The S3Config object chooses between the two, but more importantly,
implements the put_tarball() method that implements the logic to
deal with small or large objects, without pbench-backup-tarballs.py
having to worry about that.