Pass the logger when instantiating the s3 object.
We did that change earlier to the class and the instantiation of it
in pbench-backup-tarballs.py, so we don't have to pass the logger to
each and every method call that needs it. Instead, store it in the
object and retrieve it as self.logger when needed.

Do the same change in pbench-verify-backup-tarballs.py as well.
With this change, all the verify unit tests (9.*) pass - which
probably means that we don't have enough.